Better handle tests requirements, fix Travis
As long as travis does seems to have issues to run setup.py test
the previous patches needed to be patched.

Now the tests and the dev requirements are stored in extras_require
within the setup.py file.

The test directory as also been renamed to tests which is a more
consistent name accross the community